Fixed tests to make sure the message is broadcasted properly both locally and across servers.
