+++ tests/PresenceChannelTest.php
@@ -2,8 +2,12 @@
 
 namespace BeyondCode\LaravelWebSockets\Test;
 
+use BeyondCode\LaravelWebSockets\API\TriggerEvent;
 use BeyondCode\LaravelWebSockets\Server\Exceptions\InvalidSignature;
 use Carbon\Carbon;
+use GuzzleHttp\Psr7\Request;
+use Illuminate\Http\JsonResponse;
+use Pusher\Pusher;
 use Ratchet\ConnectionInterface;
 
 class PresenceChannelTest extends TestCase
@@ -454,5 +458,147 @@ class PresenceChannelTest extends TestCase
                 $this->channelManager->getRedisKey('1234', 'presence-channel'),
                 $message->getPayload(),
             ]);
+    }
+
+    public function test_it_fires_the_event_to_presence_channel()
+    {
+        $this->newPresenceConnection('presence-channel');
+
+        $connection = new Mocks\Connection;
+
+        $requestPath = '/apps/1234/events';
+
+        $routeParams = [
+            'appId' => '1234',
+        ];
+
+        $queryString = Pusher::build_auth_query_string(
+            'TestKey', 'TestSecret', 'POST', $requestPath, [
+                'name' => 'some-event',
+                'channels' => ['presence-channel'],
+                'data' => json_encode(['some-data' => 'yes']),
+            ],
+        );
+
+        $request = new Request('POST', "{$requestPath}?{$queryString}&".http_build_query($routeParams));
+
+        $controller = app(TriggerEvent::class);
+
+        $controller->onOpen($connection, $request);
+
+        /** @var JsonResponse $response */
+        $response = array_pop($connection->sentRawData);
+
+        $this->assertSame([], json_decode($response->getContent(), true));
+
+        $this->statisticsCollector
+            ->getAppStatistics('1234')
+            ->then(function ($statistic) {
+                $this->assertEquals([
+                    'peak_connections_count' => 1,
+                    'websocket_messages_count' => 1,
+                    'api_messages_count' => 1,
+                    'app_id' => '1234',
+                ], $statistic->toArray());
+            });
+    }
+
+    public function test_it_fires_event_across_servers_when_there_are_not_users_locally_for_presence_channel()
+    {
+        $connection = new Mocks\Connection;
+
+        $requestPath = '/apps/1234/events';
+
+        $routeParams = [
+            'appId' => '1234',
+        ];
+
+        $queryString = Pusher::build_auth_query_string(
+            'TestKey', 'TestSecret', 'POST', $requestPath, [
+                'name' => 'some-event',
+                'channels' => ['presence-channel'],
+                'data' => json_encode(['some-data' => 'yes']),
+            ],
+        );
+
+        $request = new Request('POST', "{$requestPath}?{$queryString}&".http_build_query($routeParams));
+
+        $controller = app(TriggerEvent::class);
+
+        $controller->onOpen($connection, $request);
+
+        /** @var JsonResponse $response */
+        $response = array_pop($connection->sentRawData);
+
+        $this->assertSame([], json_decode($response->getContent(), true));
+
+        if (method_exists($this->channelManager, 'getPublishClient')) {
+            $this->channelManager
+                ->getPublishClient()
+                ->assertCalledWithArgsCount(1, 'publish', [
+                    $this->channelManager->getRedisKey('1234', 'presence-channel'),
+                    json_encode([
+                        'event' => 'some-event',
+                        'channel' => 'presence-channel',
+                        'data' => json_encode(['some-data' => 'yes']),
+                        'appId' => '1234',
+                        'socketId' => null,
+                        'serverId' => $this->channelManager->getServerId(),
+                    ]),
+                ]);
+        }
+    }
+
+    public function test_it_fires_event_across_servers_when_there_are_users_locally_for_presence_channel()
+    {
+        $wsConnection = $this->newPresenceConnection('presence-channel');
+
+        $connection = new Mocks\Connection;
+
+        $requestPath = '/apps/1234/events';
+
+        $routeParams = [
+            'appId' => '1234',
+        ];
+
+        $queryString = Pusher::build_auth_query_string(
+            'TestKey', 'TestSecret', 'POST', $requestPath, [
+                'name' => 'some-event',
+                'channels' => ['presence-channel'],
+                'data' => json_encode(['some-data' => 'yes']),
+            ],
+        );
+
+        $request = new Request('POST', "{$requestPath}?{$queryString}&".http_build_query($routeParams));
+
+        $controller = app(TriggerEvent::class);
+
+        $controller->onOpen($connection, $request);
+
+        /** @var JsonResponse $response */
+        $response = array_pop($connection->sentRawData);
+
+        $this->assertSame([], json_decode($response->getContent(), true));
+
+        if (method_exists($this->channelManager, 'getPublishClient')) {
+            $this->channelManager
+                ->getPublishClient()
+                ->assertCalledWithArgsCount(1, 'publish', [
+                    $this->channelManager->getRedisKey('1234', 'presence-channel'),
+                    json_encode([
+                        'event' => 'some-event',
+                        'channel' => 'presence-channel',
+                        'data' => json_encode(['some-data' => 'yes']),
+                        'appId' => '1234',
+                        'socketId' => null,
+                        'serverId' => $this->channelManager->getServerId(),
+                    ]),
+                ]);
+        }
+
+        $wsConnection->assertSentEvent('some-event', [
+            'channel' => 'presence-channel',
+            'data' => json_encode(['some-data' => 'yes']),
+        ]);
     }
 }
